create data access object create database create database module in di module

+++ app/src/main/java/com/robusta/photoweather/network/ApiService.kt
@@ -9,8 +9,8 @@ interface ApiService {
 
     @GET("/data/2.5//weather")
     suspend fun getCurrentWeather(
-        @Query("lat") lat: Long,
-        @Query("lon") lon: Long,
+        @Query("lat") lat: Double,
+        @Query("lon") lon: Double,
         @Query("appid") apiKey: String,
         @Query("units") units: String? = "metric",
         @Query("lang") lang: String? = "ar",
